clase4/Practica/ml_services_comparison/graphql_service/pyproject.toml
[tool.poetry]
name = "graphql-service"
version = "0.1.0"
description = ""
authors = ["JARS29 <jandresrsalas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.200.0"}
scikit-learn = "^1.0.0" # Para el modelo toy

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"